requirements.txt
asgiref==3.9.1
autoflake==2.3.1
black==25.1.0
cfgv==3.4.0
click==8.2.1
coverage==7.11.0
distlib==0.4.0
Django==6.0
django-formtools==2.5.1
django-otp==1.7.0
django-phonenumber-field==8.4.0
django-recaptcha==4.1.0
django-two-factor-auth==1.18.1
filelock==3.19.1
gunicorn==23.0.0
identify==2.6.13
isort==6.0.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.27
platformdirs==4.3.8
pre_commit==4.3.0
psycopg2-binary==2.9.10
pyflakes==3.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
qrcode==8.2
sqlparse==0.5.3
virtualenv==20.34.0
